Use name export instead of default export.
To import the router use:

import { Router } from "@hyperapp/router"

or

const { Router } = require("@hyperapp/router")

This opens the door to adding future exports like <Link />, etc.
